Move all the routes over
