+++ src/drivers/express/ExpressAdminRouteDriver.ts
@@ -2,17 +2,12 @@ import 'dotenv/config';
 import * as express from 'express';
 import { Router } from 'express';
 import * as proxy from 'express-http-proxy';
-import * as querystring from 'querystring';
 import {
-  ADMIN_LEARNING_OBJECT_ROUTES,
   ADMIN_USER_ROUTES,
   ADMIN_MAILER_ROUTES,
 } from '../../routes';
 
 const USERS_API = process.env.USERS_API || 'localhost:4000';
-const LEARNING_OBJECT_SERVICE_URI =
-  process.env.LEARNING_OBJECT_SERVICE_URI || 'localhost:5000';
-
 
 /**
  * Serves as a factory for producing a router for the express app.rt
@@ -61,117 +56,6 @@ export default class ExpressAdminRouteDriver {
         proxyReqPathResolver: req => {
           const route = ADMIN_USER_ROUTES.DELETE_USER(req.params.id);
           return route;
-        },
-      }),
-    );
-
-    // Learning Object Routes
-    router.get(
-      '/learning-objects',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const route = `${
-            ADMIN_LEARNING_OBJECT_ROUTES.FETCH_LEARNING_OBJECTS
-          }?${querystring.stringify(req.query)}`;
-          return route;
-        },
-      }),
-    );
-    router.patch(
-      '/learning-objects',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const route = ADMIN_LEARNING_OBJECT_ROUTES.UPDATE_OBJECT();
-          return route;
-        },
-      }),
-    );
-    router.get(
-      '/learning-objects/:learningObjectId',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const route = ADMIN_LEARNING_OBJECT_ROUTES.GET_FULL_OBJECT(
-            req.params.learningObjectId,
-          );
-          return route;
-        },
-      }),
-    );
-    router.patch(
-      '/users/:username/learning-objects/:learningObjectName/publish',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectName = req.params.learningObjectName;
-          return ADMIN_LEARNING_OBJECT_ROUTES.PUBLISH_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.patch(
-      '/users/:username/learning-objects/:learningObjectName/unpublish',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectName = req.params.learningObjectName;
-          return ADMIN_LEARNING_OBJECT_ROUTES.UNPUBLISH_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.patch(
-      '/users/:username/learning-objects/:learningObjectName/lock',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectName = req.params.learningObjectName;
-          return ADMIN_LEARNING_OBJECT_ROUTES.LOCK_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.patch(
-      '/users/:username/learning-objects/:learningObjectName/unlock',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectName = req.params.learningObjectName;
-          return ADMIN_LEARNING_OBJECT_ROUTES.UNLOCK_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.delete(
-      '/users/:username/learning-objects/:learningObjectName',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectName = req.params.learningObjectName;
-          return ADMIN_LEARNING_OBJECT_ROUTES.DELETE_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.delete(
-      '/users/:username/learning-objects/multiple/:learningObjectIDs',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = req.params.username;
-          const learningObjectIDs = req.params.learningObjectIDs;
-          return ADMIN_LEARNING_OBJECT_ROUTES.DELETE_MULTIPLE_LEARNING_OBJECTS(
-            username,
-            learningObjectIDs,
-          );
         },
       }),
     );
