+++ src/drivers/express/ExpressRouteDriver.ts
@@ -5,9 +5,7 @@ import proxy = require('express-http-proxy');
 import * as querystring from 'querystring';
 import 'dotenv/config';
 import {
-  LEARNING_OBJECT_ROUTES,
   BUSINESS_CARD_ROUTES,
-  FILE_UPLOAD_ROUTES,
   STATS_ROUTE,
   ADMIN_USER_ROUTES,
   USER_ROUTES,
@@ -17,9 +15,6 @@ import {
 import { SocketInteractor } from '../../interactors/SocketInteractor';
 
 const USERS_API = process.env.USERS_API || 'localhost:4000';
-const LEARNING_OBJECT_SERVICE_URI =
-  process.env.LEARNING_OBJECT_SERVICE_URI || 'localhost:5000';
-const FILE_UPLOAD_API = process.env.FILE_UPLOAD_API || 'localhost:5100';
 const BUSINESS_CARD_API = process.env.BUSINESS_CARD_API || 'localhost:3009';
 const UTILITY_API = process.env.UTILITY_URI || 'localhost:9000';
 const COA_API = process.env.COA_SERVICE || 'localhost:8500';
@@ -67,16 +62,6 @@ export default class ExpressRouteDriver {
       }),
     );
 
-    // Retrieves the metrics for a collection
-    router.route('/:collection/metrics').get(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          console.log(req.params.collection);
-          return `/${encodeURIComponent(req.params.collection)}/metrics`;
-        },
-      }),
-    );
-
     // GUIDELINE ROLES
     router.route('/guidelines/members').get(
       proxy(USERS_API, {
@@ -102,47 +87,6 @@ export default class ExpressRouteDriver {
       }),
     );
 
-    router.route('/users/:username/learning-objects/:id')
-      .all(proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          let uri = `/users/${encodeURIComponent(req.params.username)}/learning-objects/${encodeURIComponent(req.params.id)}`;
-          if (req.query) {
-            uri += '?' + querystring.stringify(req.query);
-          }
-          return uri;
-        },
-      }));
-
-    // Retrieves the materials for a learning object
-    router.get(
-      '/users/:username/learning-objects/:id/materials',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/users/:username/learning-objects/${encodeURIComponent(req.params.id)}/materials?${querystring.stringify(
-            req.query,
-          )}`;
-        },
-      }),
-    );
-
-    router.get(
-      '/users/:username/learning-objects/:id/children',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/users/:username/learning-objects/${encodeURIComponent(req.params.id)}/children`;
-        },
-      }),
-    );
-
-    router.get(
-      '/users/:username/learning-objects/:id/parents',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/users/:username/learning-objects/${encodeURIComponent(req.params.id)}/parents`;
-        },
-      }),
-    );
-
     router.use('/users', this.buildUserRouter());
     router.use(
       '/users/:username/learning-objects',
@@ -150,48 +94,7 @@ export default class ExpressRouteDriver {
     );
 
     router.use('/learning-objects', this.buildPublicLearningObjectRouter());
-    router.get(
-      '/collections/stats',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/collections/stats`;
-        },
-      }),
-    );
-    router.get(
-      '/collections/:name',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/collections/${encodeURIComponent(req.params.name)}`;
-        },
-      }),
-    );
-    router.get(
-      '/collections/:name/meta',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/collections/${encodeURIComponent(req.params.name)}/meta`;
-        },
-      }),
-    );
-    router.get(
-      '/collections/:name/learning-objects',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/collections/${encodeURIComponent(
-            req.params.name,
-          )}/learning-objects`;
-        },
-      }),
-    );
-    router.get(
-      '/collections',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.GET_COLLECTIONS;
-        },
-      }),
-    );
+
     router.get(
       '/users/identifiers/active',
       proxy(USERS_API, {
@@ -217,17 +120,6 @@ export default class ExpressRouteDriver {
       proxy(USERS_API, {
         proxyReqPathResolver: req => {
           return `/users/password`;
-        },
-      }),
-    );
-
-    router.patch(
-      '/learning-objects/:learningObjectId/collections',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.ADD_LEARNING_OBJECT_TO_COLLECTION(
-            req.params.learningObjectId,
-          );
         },
       }),
     );
@@ -323,57 +215,6 @@ export default class ExpressRouteDriver {
         },
       }),
     );
-
-    router.post(
-      '/learning-objects/:username/:learningObjectName/children',
-      proxy(LEARNING_OBJECT_SERVICE_URI),
-    );
-    router.delete(
-      '/learning-objects/:username/:learningObjectName/children',
-      proxy(LEARNING_OBJECT_SERVICE_URI),
-    );
-
-    router.post(
-      '/learning-objects/:learningObjectId/learning-outcomes',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/learning-objects/${encodeURIComponent(
-            req.params.learningObjectId,
-          )}/learning-outcomes`;
-        },
-      }),
-    );
-    router.patch(
-      '/learning-objects/:learningObjectId/learning-outcomes/:outcomeId',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/learning-objects/${encodeURIComponent(
-            req.params.learningObjectId,
-          )}/learning-outcomes/${encodeURIComponent(req.params.outcomeId)}`;
-        },
-      }),
-    );
-    router.delete(
-      '/learning-objects/:learningObjectId/learning-outcomes/:outcomeId',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/learning-objects/${encodeURIComponent(
-            req.params.learningObjectId,
-          )}/learning-outcomes/${encodeURIComponent(req.params.outcomeId)}`;
-        },
-      }),
-    );
-    // Not in production
-    router.post(
-      '/users/:username/learning-objects/:learningObjectId/outcomes/:outcomeId/mappings',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/users/${encodeURIComponent(req.params.username)}/learning-objects/${encodeURIComponent(
-            req.params.learningObjectId,
-          )}/outcomes/${encodeURIComponent(req.params.outcomeId)}/mappings`;
-        },
-      }),
-    );
   }
 
   /**
@@ -383,30 +224,6 @@ export default class ExpressRouteDriver {
    */
   private buildUserRouter() {
     let router: Router = express.Router();
-
-    router.post(
-      '/:userId/learning-objects/:cuid/changelog',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.CREATE_CHANGELOG(
-            req.params.userId,
-            req.params.cuid,
-          );
-        },
-      }),
-    );
-    router.get(
-      '/:userId/learning-objects/:cuid/changelogs',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.GET_ALL_CHANGELOGS(
-            req.params.userId,
-            req.params.cuid,
-            req.query,
-          );
-        },
-      }),
-    );
 
     // Welcome page
     router
@@ -479,18 +296,6 @@ export default class ExpressRouteDriver {
         },
       }),
     );
-    router.all(
-      '/:userId/learning-objects/:learningObjectId/submissions',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.SUBMIT_FOR_REVIEW(
-            req.params.userId,
-            req.params.learningObjectId,
-            req.query,
-          );
-        },
-      }),
-    );
     // Remove account
     router.route('/:username').delete(
       proxy(USERS_API, {
@@ -553,41 +358,6 @@ export default class ExpressRouteDriver {
         },
       }),
     );
-
-    router.post('/:username/learning-objects/:cuid/versions', proxy(LEARNING_OBJECT_SERVICE_URI, {
-      proxyReqPathResolver: req => {
-        return `/users/${encodeURIComponent(
-          req.params.username,
-        )}/learning-objects/${encodeURIComponent(
-          req.params.cuid,
-        )}/versions`;
-      },
-    }));
-
-    router.route('/:username/learning-objects/:cuid/versions/:version/bundle').get(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/users/${encodeURIComponent(
-            req.params.username,
-          )}/learning-objects/${encodeURIComponent(
-            req.params.cuid,
-          )}/versions/${encodeURIComponent(
-            req.params.version,
-          )}/bundle`;
-        },
-      }),
-    );
-
-    // TODO: Update route in client
-    router.get('/:username/learning-objects/:learningObjectId/outcomes', proxy(LEARNING_OBJECT_SERVICE_URI, {
-      proxyReqPathResolver: req => {
-        return `/users/${encodeURIComponent(
-          req.params.username,
-        )}/learning-objects/${encodeURIComponent(
-          req.params.learningObjectId,
-        )}/outcomes`;
-      },
-    }));
 
     router.get(
       '/search',
@@ -660,298 +430,6 @@ export default class ExpressRouteDriver {
   ) {
     let router: Router = express.Router({ mergeParams: true });
     const parentParams = _req.params;
-    router
-      .route('')
-      .get(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            return (
-              LEARNING_OBJECT_ROUTES.LOAD_LEARNING_OBJECT_SUMMARY(
-                _req.params.username,
-              ) +
-              '?' +
-              querystring.stringify(req.query)
-            );
-          },
-        }),
-      )
-      .post(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const authorUsername = parentParams.username;
-            return LEARNING_OBJECT_ROUTES.CREATE_LEARNING_OBJECT(
-              authorUsername,
-            );
-          },
-        }),
-      );
-
-    router.delete(
-      '/:learningObjectName',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const learningObjectName = req.params.learningObjectName;
-          return LEARNING_OBJECT_ROUTES.DELETE_LEARNING_OBJECT_BY_NAME(
-            learningObjectName,
-          );
-        },
-      }),
-    );
-
-    router
-      .route('/:learningObjectId')
-      .patch(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const authorUsername = parentParams.username;
-            const learningObjectId = req.params.learningObjectId;
-            return LEARNING_OBJECT_ROUTES.UPDATE_LEARNING_OBJECT({
-              authorUsername,
-              id: learningObjectId,
-            });
-          },
-        }),
-      )
-      .delete(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const authorUsername = parentParams.username;
-            const learningObjectId = req.params.learningObjectId;
-            return LEARNING_OBJECT_ROUTES.DELETE_LEARNING_OBJECT({
-              authorUsername,
-              id: learningObjectId,
-            });
-          },
-        }),
-      );
-
-    router.get(
-      '/profile',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.LOAD_USER_PROFILE(
-            encodeURIComponent(req.params.username),
-          );
-        },
-      }),
-    );
-    router.patch(
-      '/:learningObjectName/publish',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.PUBLISH_LEARNING_OBJECT;
-        },
-      }),
-    );
-    router.patch(
-      '/:learningObjectName/unpublish',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.UNPUBLISH_LEARNING_OBJECT;
-        },
-      }),
-    );
-    router.delete(
-      '/multiple/:names',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          let names = req.params.names.split(',');
-          return LEARNING_OBJECT_ROUTES.DELETE_MULTIPLE_LEARNING_OBJECTS(names);
-        },
-      }),
-    );
-
-    router.post(
-      '/:learningObjectId/revisions',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.CREATE_LEARNING_OBJECT_REVISION(
-            req.params.username,
-            req.params.learningObjectId,
-          );
-        },
-      }),
-    );
-
-    router.get(
-      '/:learningObjectId/revisions/:revisionId',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.GET_LEARNING_OBJECT_REVISION({
-            username: req.params.username,
-            learningObjectId: req.params.learningObjectId,
-            revisionId: req.params.revisionId,
-            query: req.query,
-          });
-        },
-      }),
-    );
-    // FILE OPERATIONS
-    /**
-     * TODO: Deprecate in favor of more RESTful `/:learningObjectId/materials/files/:fileId` when clients have updated
-     */
-    router
-      .route('/:learningObjectID/files/:fileId')
-      .patch(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            const learningObjectId = req.params.learningObjectID;
-            const fileId = req.params.fileId;
-            return LEARNING_OBJECT_ROUTES.UPDATE_FILE({
-              username,
-              learningObjectId,
-              fileId,
-            });
-          },
-        }),
-      )
-      .delete(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            const learningObjectId = req.params.learningObjectID;
-            const fileId = req.params.fileId;
-            return LEARNING_OBJECT_ROUTES.UPDATE_FILE({
-              username,
-              learningObjectId,
-              fileId,
-            });
-          },
-        }),
-      );
-    router
-      .route('/:learningObjectId/materials/files/:fileId')
-      .patch(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            const learningObjectId = req.params.learningObjectId;
-            const fileId = req.params.fileId;
-            return LEARNING_OBJECT_ROUTES.UPDATE_FILE({
-              username,
-              learningObjectId,
-              fileId,
-            });
-          },
-        }),
-      )
-      .delete(
-        proxy(LEARNING_OBJECT_SERVICE_URI, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            const learningObjectId = req.params.learningObjectId;
-            const fileId = req.params.fileId;
-            return LEARNING_OBJECT_ROUTES.UPDATE_FILE({
-              username,
-              learningObjectId,
-              fileId,
-            });
-          },
-        }),
-      );
-    router.patch(
-      '/:id/pdf',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req =>
-          LEARNING_OBJECT_ROUTES.UPDATE_PDF(req.params.id),
-      }),
-    );
-    router.get(
-      '/:id/files/:fileId/download',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = parentParams.username;
-          return LEARNING_OBJECT_ROUTES.DOWNLOAD_FILE({
-            username,
-            id: req.params.id,
-            fileId: req.params.fileId,
-            query: req.query,
-          });
-        },
-      }),
-    );
-    router.route('/:id/materials').get(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = parentParams.username;
-          const id = req.params.id;
-          return LEARNING_OBJECT_ROUTES.GET_MATERIALS({
-            username,
-            id,
-            query: req.query,
-          });
-        },
-      }),
-    );
-    router.route('/:id/materials/files').post(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = parentParams.username;
-          const id = req.params.id;
-          return LEARNING_OBJECT_ROUTES.ADD_MATERIALS(username, id);
-        },
-      }),
-    );
-    /**
-     * FIXME: This route should be removed when the API is tested and  client is updated
-     */
-    router.route('/:objectId/files/:fileId/multipart').all(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          const username = parentParams.username;
-          return FILE_UPLOAD_ROUTES.INIT_MULTIPART({
-            username,
-            objectId: req.params.objectId,
-            fileId: req.params.fileId,
-          });
-        },
-      }),
-    );
-
-    /**
-     * FIXME: The admin suffix should be remove when API is tested and client is updated
-     */
-    router.route('/:objectId/files/:fileId/multipart/admin').post(
-      proxy(FILE_UPLOAD_API, {
-        proxyReqPathResolver: req => {
-          const username = parentParams.username;
-          return FILE_UPLOAD_ROUTES.INIT_MULTIPART({
-            username,
-            objectId: req.params.objectId,
-            fileId: req.params.fileId,
-          });
-        },
-      }),
-    );
-    router
-      .route('/:objectId/files/:fileId/multipart/:uploadId/admin')
-      .patch(
-        proxy(FILE_UPLOAD_API, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            return FILE_UPLOAD_ROUTES.FINALIZE_MULTIPART({
-              username,
-              objectId: req.params.objectId,
-              fileId: req.params.fileId,
-              uploadId: req.params.uploadId,
-            });
-          },
-        }),
-      )
-      .delete(
-        proxy(FILE_UPLOAD_API, {
-          proxyReqPathResolver: req => {
-            const username = parentParams.username;
-            return FILE_UPLOAD_ROUTES.ABORT_MULTIPART({
-              username,
-              objectId: req.params.objectId,
-              fileId: req.params.fileId,
-              uploadId: req.params.uploadId,
-            });
-          },
-        }),
-      );
     return router(_req, res, next);
   }
 
@@ -964,70 +442,7 @@ export default class ExpressRouteDriver {
    */
   private buildPublicLearningObjectRouter() {
     let router: Router = express.Router();
-
-    router.get(
-      '',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          let route = `${
-            LEARNING_OBJECT_ROUTES.FETCH_LEARNING_OBJECTS
-            }?${querystring.stringify(req.query)}`;
-          return route;
-        },
-      }),
-    );
-    router.route('/:learningObjectId').get(
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return `/learning-objects/${encodeURIComponent(
-            req.params.learningObjectId,
-          )}`;
-        },
-      }),
-    );
-    router.get(
-      '/stats',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return STATS_ROUTE.LEARNING_OBJECT_STATS;
-        },
-      }),
-    );
-    router.get(
-      '/:author/:learningObjectName',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          let username = req.params.author;
-          let learningObjectName = req.params.learningObjectName;
-          return LEARNING_OBJECT_ROUTES.LOAD_LEARNING_OBJECT(
-            username,
-            learningObjectName,
-          );
-        },
-      }),
-    );
-    router.get(
-      '/:author',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.FETCH_USERS_LEARNING_OBJECTS(
-            req.params.author,
-          );
-        },
-      }),
-    );
-    router.get(
-      '/:id/children/summary',
-      proxy(LEARNING_OBJECT_SERVICE_URI, {
-        proxyReqPathResolver: req => {
-          return LEARNING_OBJECT_ROUTES.GET_LEARNING_OBJECT_CHILDREN(
-            req.params.id,
-          );
-        },
-      }),
-    );
-
-
+    
     return router;
   }
 }
